requirements.txt
fastapi[standard]==0.112.2
PyJWT[crypto]==2.9.0
requests==2.32.3
uvicorn==0.30.6
pydantic==2.9.2
Jinja2==3.1.4
autopep8==2.3.1
python-slugify==8.0.4